requirements.txt
streamlit
langchain
langchain-community
torch
sentence-transformers
pypdf
chromadb-client
transformers
accelerate
bitsandbytes
langchain-unstructured[local]